strategy works with artboard and layer for move down
